Rename model to scores to avoid name conflict
Since we're naming the genserver that processes users spinner already, I
needed a new name for the model. It's technically just storing the users
score so that's what we're calling it now, score.

We might not even need this so much to be honest. We'll see where we go
with team score keeping I guess.