Implement spinning
The channel now spins to send client update on the state of the spinner.
The spinner genserver is now also spinning updating its score. The time
is currently arbitrary just to show a working concept. The code also
needs a little clean up.
This will do for now.
Next is documentation and figure out how to write tests for genserver.