Makes spinner_pid "the" user id for now.
Since both arena channel and squad channel pretty much relies on spinner
pid to be there for references/look ups, we're making it the user id for
now. This way,we don't have to worry about generating one for user just
yet.
This is obviously not a great long-term solution, since we won't be able
to do spectator channel without also assigning them as spinner. Plus,
pid as user id probably won't work out well in a distributed system and
make unit testing hard. We will eventually needs to come up with a
different plan.