Pass spinner struct to introspect
The introspect should be getting the spinner struct (GenServer state)
instead of what was assigned in the socket since those are going to
contained some internal information as well.
The cleaning and filtering of data to be passed to client should be done
in the Channel layer as well, not the GenServer.