Add a cleaned-up member listing to spinning
In each spin, we now take the updated member data and build a list of
members with data we intend to pass to clients. Currently, they are set
up to be sorted by connected_at datetime thought that doesn't quite work
yet. Nonetheless, we now have a clean list we build during each spinning
cycle that we can broadcast to clients without rebuilding for each
client.