Add comments and test cases.
Currently the test for the `Gyro.Squad` module is splitted into two test
files just to make it easier. One tests the portion that works with the
socket, and theo there tests the underlying GenServer implementation.
Eventually, we will need to fix this, hopefully by splitting the
GenServer code itself into 2 pieces.
Also fix an error found while writing unit tests.
Need to be better with keeping up with comments and testing.