Refactor how we test for squad members
Instead of using the `introspect` method from the Squad module, we are
now using the GenServer module to call and fetch the state, then iterate
through to find member based on `spinner_pid` stored in the socket.
The reason being that we do not want to test the `introspect` method or
rely on what information is being passed down to the socket. With
this, we eliminate those and directly test from the state of the Squad
GenServer which guarantees to have member list.

Also add support for the tag `:skip` to be excluded when we don't want
to run certain tests.