Make sure to leave squad channel before join
If we already joined a sqaud channel, we should leave before joining
another.
Eventually, we should make the server-side check for this as well.
However, it's fairly non-trivial to implement. Seems like there is
already support for this in the current master branch of Phoenix so we
will just wait for that.