Add Agents to Arena GenServer process.
Start 2 Agents when start the Arena GenServer for holding the list of
the spinners and list of squads active in the system. Because the list
could be large and doesn't need to be recalculate, it's better to put it
separately from the GenServer process itself.
This should keep the state like and make the spinning a little less
intensive.