Start a new Squad under the Application Supervisor
Instead of starting a Squad, linked to the spinner process that
requested it, we should be linking it to a different Supervisor. Linking
the Squad to a spinner process will cause it to terminate when that
spinner leaves.
However, creating an entirely new Supervisor just for the Squads seems a
little overkill. For now, we're just asking the Application
`Gyro.Supervisor` to supervise the squads for us.