Pattern matching on no process error
Apparently the GenServer could exit between checking if it exists and
calling for `introspect`. The better way to handle this is to actually
catch for response from `introspect`. If it comes back with {:noproc,_}
message then we can handle that the spinner isn't there any more and
this is probably more solid way to approach this.