Refactor Squad's members to use Map
Instead of list, we're using Map instead which allows for quicker look
up for when we need to delist. It also ensures we don't have duplicate
member twice with the same pid.
We also change to store only just the members' pids, but not their
state. Since we need to update every members in every spin anyway,
there's little value in knowing their states. Instead, this will make
the Squad state itself a lot smaller since we're storing less
information.