Spinner's introspect catch more generic :exit
The Spinner's introspect function should also catch a more generic
`:exit` message. There are cases when a spinner call may not return the
exact `:noproc` messsage but we still want to return `nil` back anyway
with the same assumption that the spinner no longer exists. This should
stop the error message we're seeing in stress testing.