Switch enlist and delist in Squad to cast
Similar to what we did with Arena, making enlisting and delisting
process in the squad a cast means it will be called asynchronously. This
should avoid timeout issue when having a lot of people joing the same
team at once. We also don't really need the state return as a result of
changing the member list everytime either.
When that is the case, we should explicitly call for the introspect
instead.