Switch Squad and Arena away from `send_interval`
Since it's not super critical that we have to perform the update at the
interval, it would reduce the workload if we just send the message to do
it again after a period instead. This means we won't have to worry about
the time it takes to complete a spin is longer than the interval calling
spin function.