Add/remove name of squad when joining/leaving
Make sure we made the information about which squad a spinner belongs to
available in the spinner state itself. We should have this info save
somewhere I supposed.
Also make sure we remove the spinner pid before we broadcast this over
the socket as well since it won't be encode very well.