Delist spinner from old squad first
Before enlisting a spinner to a squad, we check if they belong to a
squad and delist them first. That way, they can't join multiple squads.
The current version of the code doesn't unsubscribe you from the channel
yet. And neither check if you're rejoining the same squad again. It will
still go through the procedure of delisting and enlisting to the same
squad regardless. This could potentially be improved in the future.