Add legendary spinners to scoreboard.
To identify one spinner from another, an id field is needed in the
spinner state which for now is just a pid converted to string.

The legendary board is currently not detecting inactive spinners, hence,
not setting the SPM to 0 either. That means it will show up spinning on
the front-end as if the spinner is still accumulating points.

the algorithm for building the legendary board could use a little clean
up as well.