Update phoenix and phoenix_ecto packages.
This will allow us to upgrade to Elixir 1.3, Erlang OTP 19, as well as
the newer poison package which has functionality that allows us to
dictate which field in a datastruct to JSONify. This will help clean up
some of our custom code for removing data type like pid or datetime.