Create arena for squads.
The squad arena will now track which squads are in the system and using
scoreboard to also track how they perform. To get this to work, I have
to implement the `id` field in the squad struct since this is how it's
identifying duplicate entry. This actually ends up giving a field to
store the pid name and the name of the squad separately anyway so it's a
good thing to have regardless.